requirements.txt
asgiref==3.2.7
attrs==19.3.0
certifi==2020.4.5.2
chardet==3.0.4
Django==3.0.6
django-admin==2.0.1
django-environ==0.4.5
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-filter==2.3.0
django-rest-auth==0.9.5
django-six==1.0.4
djangorestframework==3.11.0
Faker==4.1.1
idna==2.9
marshmallow==3.6.0
multidict==4.7.6
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
screen==1.0.1
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
unicorn==1.0.1
urllib3==1.25.9
xlwt==1.3.0
yarl==1.4.2
